+++ TFD_Objects/Classes/Flags.hpp
@@ -1,0 +1,164 @@
+class TFD_DrapeauUE : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (UE)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_UE.paa'"
+	}
+};
+
+class TFD_DrapeauFR : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (France)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_FR.paa'"
+	}
+};
+
+class TFD_DrapeauBE : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (Belgique)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_BE.paa'"
+	}
+};
+
+class TFD_DrapeauTFP : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (TFP)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_TFP.paa'"
+	}
+};
+
+class TFD_Drapeau24eBAC : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (24eBAC)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_24eBAC.paa'"
+	}
+};
+
+class TFD_DrapeauTFD : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (TFD)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_TFD.paa'"
+	}
+};
+
+class TFD_DrapeauURSS : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (URSS)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_URSS.paa'"
+	}
+};
+
+class TFD_DrapeauOPEXKoto : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (OPEX Suricate)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_OPEXKoto.paa'"
+	}
+};
+
+class TFD_DrapeauCH : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (Republique Populaire de Chine)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_CH.paa'"
+	}
+};
+
+class TFD_DrapeauCHPLA : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Etendard (Republique Populaire de Chine)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_CHPLA.paa'"
+	}
+};
+
+class TFD_DrapeauGB : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (Gabon)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_GB.paa'"
+	}
+};
+
+class TFD_DrapeauJP : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (Japon)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_JP.paa'"
+	}
+};
+
+class TFD_DrapeauJPJSDF : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Etendard (Japon)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_JPJSDF.paa'"
+	}
+};
+
+class TFD_DrapeauMO : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+    scope = 2;
+	scopeCurator=2;		
+    displayName="Drapeau (Maroc)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_MO.paa'"
+	}
+};
+
+class TFD_DrapeauNG : FlagCarrier {
+	_generalMacro = "FlagCarrier"; 
+	scope = 2;
+	scopeCurator=2;		
+	displayName="Drapeau (Niger)";
+	author = "Heartbroken";	
+	class EventHandlers {
+		init="(_this select 0) setFlagTexture '\TFD_Objects\Data\Drapeau_NG.paa'"
+	}
+};
