+++ TFD_Objects/Classes/SignAdSponsor.hpp
@@ -1,0 +1,161 @@
+class TFD_PanneauVigipirate : SignAd_SponsorS_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (Vigipirate)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_Vigipirate.paa"};
+};
+
+class TFD_PanneauVigipirateAA : SignAd_SponsorS_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (Vigipirate Alerte Attentat)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_VigipirateAA.paa"};
+};
+
+class TFD_PanneauTerMil : SignAd_SponsorS_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (Terrain Militaire)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_TerMil.paa"};
+};
+
+class TFD_PanneauFOBYork : SignAd_SponsorS_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (FOB York)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_FOBYork.paa"};
+};
+
+class TFD_PanneauFOBMangomak : SignAd_SponsorS_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (FOB Mangomak)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_FOBMangomak.paa"};
+};
+
+class TFD_PanneauCampNala : SignAd_SponsorS_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (Camp Nala)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_CampNala.paa"};
+};
+
+class TFD_PanneauPPS : SignAd_SponsorS_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (PPS)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_PPS.paa"};
+};
+
+class TFD_PanneauPPR : SignAd_SponsorS_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (PPR)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_PPR.paa"};
+};
+
+class TFD_PanneauPPM : SignAd_SponsorS_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (PPM)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_PPM.paa"};
+};
+
+class TFD_PanneauZDHS : SignAd_SponsorS_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (ZDHS)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_ZDHS.paa"};
+};
+
+class SignAd_Sponsor_F;
+
+class TFD_PanneauCampTimon : SignAd_Sponsor_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (Camp Timon)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_CampTimon.paa"};
+};
+
+class TFD_PanneauCampPumba : SignAd_Sponsor_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (Camp Pumba)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_CampPumba.paa"};
+};
+
+class TFD_PanneauCampHanovre : SignAd_Sponsor_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (Camp Hanovre)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_CampHanovre.paa"};
+};
+
+class TFD_PanneauEnglishStadium : SignAd_Sponsor_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (English's Stadium)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_EnglishStadium.paa"};
+};
+
+class TFD_PanneauCheckpoint : SignAd_Sponsor_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (Checkpoint)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_Checkpoint.paa"};
+};
+
+class TFD_PanneauEntreeBase : SignAd_Sponsor_F {
+	_generalMacro = "SignAd_SponsorS_F";
+	scope = 2; 
+	scopeCurator = 2;	
+	displayName="Panneau (Entree Base)";
+	author = "Heartbroken";
+	hiddenSelections[]= {"camo"};
+	hiddenSelectionsTextures[]= {"\TFD_Objects\Data\Panneau_EntreeBase.paa"};
+};
